First version of Project Pathfinder
First upload of Project Pathfinder code. In this version, an ESP8266 NodeMCU is connected through I2C with an Arduino Mega (master & slave, respectively). The ESP8266 receives data through MQTT (Beebotte is used as a broker) and it's transmitted to the Arduino. The Arduino then acts accordingly.

In this release, two commands are available:

- "direction": Triggers the directionControl() function, which sets the motor (only a single motor is used in this version) at full speed for two seconds, then reverses its direction for another two seconds.

- "speed": Triggers the speedControl() functions, which accelerates the motor from zero to maximum speed and then decelerates again to zero.