First commit of the app
First commit of the Android app used to control the robot. The first version consists of a simple joystick that follows the user on touch and returns how much the joystick was pushed in cartesian coordinates (X and Y). The only files included here are the JoystickView and MainActivity (Java).

Big thanks to EfficientIsoceles for the Instructable "A Simple Android UI Joystick" which I used as my main source of inspiration.